A few minor modifications. Now uses parseFloat instead of parseInt for the lessThan and greaterThan predicates.
git-svn-id: http://greenlight.googlecode.com/svn/trunk@14 018bb2bb-3d32-9f68-d2a2-e3763db289ef

+++ GreenLight.js
@@ -213,7 +213,7 @@ GreenLight.core.__init__ = function (GreenLight, undefined) {
         lessThan: function (value) {
             if (typeof value !== "number") throw ("Argument must be a Number.");
             return function (e) {
-                var eVal = parseInt(e.value, 10);
+                var eVal = parseFloat(e.value, 10);
                 return eVal < value;
             };
         },
@@ -222,7 +222,7 @@ GreenLight.core.__init__ = function (GreenLight, undefined) {
         greaterThan: function (value) {
             if (typeof value !== "number") throw ("Argument must be a Number.");
             return function (e) {
-                var eVal = parseInt(e.value, 10);
+                var eVal = parseFloat(e.value, 10);
                 return eVal > value;
             };
         },
